PROGRESS! But still broken code (6 SAAT OLDU DELİRDİM)
