+++ app/build.gradle
@@ -9,7 +9,7 @@ android {
 
     defaultConfig {
         applicationId "com.example.p2pchat"
-        minSdk 31
+        minSdk 28
         targetSdk 32
         versionCode 1
         versionName "1.0"
